+++ web/package.json
@@ -8,7 +8,6 @@
   },
   "dependencies": {
     "@popperjs/core": "^2.10.2",
-    "@vue/eslint-config-airbnb": "^5.3.0",
     "axios": "^0.21.2",
     "bootstrap": "^5.1.1",
     "copy-to-clipboard": "^3.3.1",
@@ -29,20 +28,11 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/vue": "^6.4.2",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
     "@vue/cli-plugin-babel": "~4.5.13",
-    "@vue/cli-plugin-eslint": "~4.5.13",
     "@vue/cli-plugin-router": "~4.5.13",
     "@vue/cli-plugin-typescript": "~4.5.13",
     "@vue/cli-service": "~4.5.13",
     "@vue/compiler-sfc": "^3.2.1",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-vue": "^7.15.2",
     "husky": "4",
     "node-sass": "^5.0.0",
     "prettier": "^2.3.2",
