+++ web/src/main.ts
@@ -2,7 +2,7 @@ import { createApp } from 'vue';
 import firebase from 'firebase/compat/app';
 import router from './router';
 
-import App from '@/App.vue'
+import App from './App.vue'
