+++ web/src/shims-vue.d.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 declare module '*.vue' {
   import type { DefineComponent } from 'vue'
   const component: DefineComponent<{}, {}, any>
